refactor: organize API into modules
